Update and rename tests
